scholarship_scraper/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
celery[redis]
instaloader
requests
playwright
beautifulsoup4
lxml
python-dateutil
yt-dlp
openai-whisper
